JumpGame
You are given an integer array nums. You are initially positioned at the array's first index, and each element in the array represents your maximum jump length at that position.

Return true if you can reach the last index, or false otherwise.



Solution:

The intuition behind the code is to use a greedy approach to determine if it is possible to reach the last index of the array nums using the maximum jump length at each position. The idea is to iteratively update the maximum reachable index while traversing the array from left to right. If the maximum reachable index ever reaches or exceeds the last index, it means we can reach the end.