SearchInRotatedSortedArray
There is an integer array nums sorted in ascending order (with distinct values).

Prior to being passed to your function, nums is possibly rotated at an unknown pivot index k (1 <= k < nums.length) such that the resulting array is [nums[k], nums[k+1], ..., nums[n-1], nums[0], nums[1], ..., nums[k-1]] (0-indexed). For example, [0,1,2,4,5,6,7] might be rotated at pivot index 3 and become [4,5,6,7,0,1,2].

Given the array nums after the possible rotation and an integer target, return the index of target if it is in nums, or -1 if it is not in nums.

You must write an algorithm with O(log n) runtime complexity.

 


Solution:
At any point during the search in the rotated array, one half (either the left or the right) will always be sorted. Determining which half is sorted is crucial for our modified binary search.

If left half [low...mid][low...mid][low...mid] is sorted: We know this if the element at low\text{low}low is less than or equal to the element at mid\text{mid}mid. In a normally sorted array, if the start is less than or equal to the midpoint, it means all elements till the midpoint are in the correct increasing order.

If the target lies within this sorted left half: We know this if the target is greater than or equal to the element at low\text{low}low and less than the element at mid\text{mid}mid. If this is the case, we then move our search to this half, meaning, we update high\text{high}high to mid−1\text{mid} - 1mid−1.

Otherwise: The target must be in the right half. So, we update low\text{low}low to mid+1\text{mid} + 1mid+1.

If right half [mid...high][mid...high][mid...high] is sorted: This is the else part. If the left half isn't sorted, the right half must be!

If the target lies within this sorted right half: We know this if the target is greater than the element at mid\text{mid}mid and less than or equal to the element at high\text{high}high. If so, we move our search to this half by updating low\text{low}low to mid+1\text{mid} + 1mid+1.

Otherwise: The target must be in the left half. So, we update high\text{high}high to mid−1\text{mid} - 1mid−1