ArrangementArrayBasedOnSecondArray
﻿
You are given two array A[] and B[]. Your task is to arrange elements of first array according to the order defined by second array. The first array contains integers which represent heights of persons and second array defines how many persons in front of him should stand with height greater than or equal to that person.
Input:
First line of input contains number of testcases T. For each testcase, there will be three lines, first of which will contain an integer N denoting the size of the array. Second line contains N space separated integers for array A[] denoting the height of the persons and last line of the test case contains the arrangement array B[].
Output:
Print the result array showing the new height arrangement.
Your Task:
You don't need to read input or print output. Your task is to complete the function arrange() that takes array A, array B, integer N as input parameters and returns a list of integers denoting the new arrangement of people according to their heights. Constraints:
1 <= T <= 50
1 <= N<=103
1 <= A[i] <= 104
1 <=B] <= N
Example:
Sample Input:
2
3
321
011
6
532614
012032
Sample Output:
312
532164
Explanation:
Testcase 1:
The output is:
312
Corresponding B value: 011
O people of greater or equal height in front of person of height 3.
1 person of greater or equal height in front of person of height 1.
1 person of greater or equal height in front of person of height 2.