Create BinaryTreeQn.java
﻿
There is a computer network connected as a binary tree. An engineer want to connect any of the leaf computers with all other leaf computer with win wire. It is not possible to connect two leaf computer if the difference between their depth from the root is greater than 1. You need to tell your Boss whether it is possible to connect any of the leaf
computers with all other leaf computers
Note that, if the network contain only one leaf computer then you don't need to connect it. That is, the output will be "YES"
Input:
First line of input contains the number of test cases T. For each test case, there will be only a single line of input which is a string representing the network as described below
1. The values in the string are in the order of verval of the network where number decomputer, and a character "dent NULL child
2. For exampl
For the above tree, the string will be 123N N46NSNN7N Output:
For each test case, print "Yes" (without quotes) above mentioned connection is possible otherwise, print "No" (without quote in new Your Task
Your task is to complete the function Possible) which takes root pointer as parameter and return valu
Constrainti
1 T 100
1 cm Number of computers 10000
Example
Input:
2
1234567
12345NNG
Output:
Yo
No
Explanation
Test Case 1:
All leaf computer at all. Test Case 2:
6
4
5
3
Leaf computers are 3,5&6 and their depth are 2.3 and 4 respectively and (4-2)-2>1.