DP Recursion
﻿
There are N Algorithm Experts and N-1 Software Developers. You as Project manager will exclude an Algorithm Expert and form N-1 teams each consist of a software developer and an algorithm expert such that the efficiency of the whole project is maximised, but the problem is you can't jumble the ordering as the employees won't be happy.
There is an array algoExperts[] which denotes the skill of ith algorithm expert and an array Developers[] which denotes the skill of ith software developer.
Efficency of a team is given by the product of the skill of software developer and the algorithm expert, and total efficiency is the sum of the efficiencies of N-1 teams.
Input:
First line will contain an integer T denoting the number of testcases. First line of each testcase contains a single integer N, denoting the size of the array. Next line conatains N space separated integers denoting the skill of the Algorithm Experts and last line of each teastcases contains N-1 elements denoting the skills of N-1 Software Developers.
Output:
Print the maximum total efficiency of the team formed.
Your Task:
You don't need to take input. Complete the function project() which takes algoExperts[]. Developers[] and size n as input and return the maximum efficiency of the formed team. Printing is done by the driver code only. Constraints:
1<=T<=30
2<=N<=105
O<algoExperts[i]<=106
0<=Developers[i]<=106
Example:
Input:
2
5
41529
1211
2
41
1
Output:
25
4
Explanation:
Testcase 1: 2nd Algorithm Expert is excluded and the maximum efficiency will be 4*1 +5°2+2*1+9*1=25.
Testcase 2: 2nd Algorithm Expert is excluded and the maximum efficiency will be 4*1 = 4.