ReminderOfFraction
﻿
Ted is a pro coder, but today he is very busy in doing college assignments. He needs your help to complete a task. The task is as follow:-
Given two positive integers A and B denoting numerator and denomenator of a fraction. You need to count the length of repeated pattern of digits in decimal part of A/B. Input:
First line of input contains an integer T denoting number of test cases. For each test case, there will be two space seperated positive integers A and B.
Output:
For each test case print the length of repeated pattern of digits in decimal part of A/B.
Task:
This is a functional problem, you don't need to take input just complete the function findCount() which accepts two integers A and B as parameters and return the length of repeated pattern of digits in decimal part of A/B.
Constraints:
1 <=T<= 100
1 <= A, B <= 10000
Example:
Input:
3
12
227
84
Output:
0
6
0
Explanation:
Testcase1: 1/2 = 0.5 since there is no repetition in decimal part of 1/2 therefore ans is 0.
Testcase2: 22/7 = 3.142857142857142857.... length of repeating pattern in decimal part is 6. Testcase3: 8/4 = 2, since 4 completly divides 8 there is no decimal part.